+++ packages/core/src/2d/sprite/SpriteRenderer.ts
@@ -3,9 +3,8 @@ import { Camera } from "../../Camera";
 import { assignmentClone, deepClone, ignoreClone } from "../../clone/CloneManager";
 import { ICustomClone } from "../../clone/ComponentCloner";
 import { Entity } from "../../Entity";
-import { RendererUpdateFlags } from "../../enums/RendererUpdateFlags";
 import { ListenerUpdateFlag } from "../../ListenerUpdateFlag";
-import { Renderer } from "../../Renderer";
+import { Renderer, RendererUpdateFlags } from "../../Renderer";
 import { CompareFunction } from "../../shader/enums/CompareFunction";
 import { Shader } from "../../shader/Shader";
 import { ShaderProperty } from "../../shader/ShaderProperty";
