--- packages/core/src/enums/RendererUpdateFlags.ts
@@ -1,9 +1,0 @@
-/**
- * @internal
- */
-export enum RendererUpdateFlags {
-  /** None. */
-  None = 0,
-  /** Include world position and world bounds. */
-  WorldVolume = 0x1
-}
